Add checkFirst and checkLast functions
These functions will verify if the first character of the string is
an operand, in which case it'll be deleted. Will also check if the user
is trying to input two operands together, which would obviously break
the expression, then, if that's the case it'll delete the last
operand and input the new one.